requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
drf-yasg>=1.21
types-python-dateutil
python-dateutil
google-genai
django-cors-headers

gunicorn>=21.2

psycopg[binary]>=3.1   # PostgreSQL (for deployment)

whitenoise>=6.6        # static files in production

python-dotenv>=1.0     # optional (env variables)
pytest-django